requirements.txt
aiogram>=3.10.0
asyncpg
redis
pandas
openpyxl
pyxlsb
python-dotenv
loguru
aiohttp
matplotlib
numpy
scipy
scikit-learn